Add docker container for development
Add Dockerfile for development container for project. This image
is based on zephyr-build container for all the development utilities
for zephyr but also adds project-specific utilities.

Add Task support to the development environment.
"Task is a task runner / build tool that aims to be simpler and easier
to use than, for example, GNU Make."

Add VSCode devcontainer support based on this container - devcontainer.json
file.