Fix application structure and build problems
There were problems with building Zephyr samples from
west workspace due to misconfigration of application structure.
Now properly use Zephyr's T2 start topology based on example-application
repository.
This structure looks like this:
The top level repository is the repository of the project. It is one
level above the manifest repository to include the zephyr workspace,
since we also want to keep other files in this level.
|- .devcontainer/
|- taskfiles/
|- .vscode/
|- controller/ - manifest repository based on example-application.
  |- app/ - application sources
    |- prj.conf
    |- src/
  |- drivers/ - zephyr drivers
  |- lib/ - zephyr libraries
  |- boards/ - zephyr boards